+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## [1.0.1-beta] - 2024-09-02
+**Added**
+- Added interactive mode for stack selection to `apply`, `delete`, `show` commands
+- Added `--all/-A` flag to `apply` and `delete` commands to select all stacks
+
+**Updated**
+- Updated aws_sdk crate verisons
 
 ## [1.0.0-beta] - 2024-08-24
 **Added**